Use a dedicated user to run the server
Let systemd create the storage directory and use a dedicated
'passim' user for the service.

PrivateUsers=yes is not the ideal solution. It uses a user namespace,
but the service still runs as root in the namespace. Firstly, user namespaces
are not available in all contexts, in particular they may simply be
disabled, because in the past that interface had a lot of vulnerabilities
and some distros or local policies disable it. Secondly, even if the user
namespace is available, the exposure from running as root in a user namespace
is still much higher than running as an unprivileged user. (User namespaces
are in general used exactly for this purpose: we can pretend to be root
and do stuff we wouldn't be able to do otherwise.)  We can get a more
comprehensive and robust isolation if we don't run as root at all.