Do not abort with a critical warning when no query is used
It's also worth noting that the "crash on anything looking weird" is 100%
deliberate -- i.e. we're doing `G_DEBUG=fatal-criticals` so that any critical
warning causes an instant "crash" -- on the logic a dead daemon is better than
a buffer or stack overflow.

Fixes https://github.com/hughsie/passim/issues/25